Init Commit
Added new skybox textures and fixed the upside down skybox. Added an easy way to change the current skybox as well. Shadow maps now follow a sun/moon orbit and will switch between the library and boss area depending on which scene the player is in. Fixed jittering and jagged shadow maps